# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Interior Design Studio
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectPhase = typeof ProjectPhase[keyof typeof ProjectPhase];


export const ProjectPhase = {
  past: 'past',
  present: 'present',
  future: 'future',
} as const;

export interface Project {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  phase: ProjectPhase;
  location: string;
  /** @nullable */
  style?: string | null;
  /** @nullable */
  area?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  completionYear?: number | null;
  /** @nullable */
  clientName?: string | null;
  featured?: boolean;
  createdAt: string;
}

export type ProjectInputPhase = typeof ProjectInputPhase[keyof typeof ProjectInputPhase];


export const ProjectInputPhase = {
  past: 'past',
  present: 'present',
  future: 'future',
} as const;

export interface ProjectInput {
  title: string;
  description?: string;
  phase: ProjectInputPhase;
  location: string;
  style?: string;
  area?: string;
  imageUrl?: string;
  completionYear?: number;
  clientName?: string;
  featured?: boolean;
}

export type ProjectUpdatePhase = typeof ProjectUpdatePhase[keyof typeof ProjectUpdatePhase];


export const ProjectUpdatePhase = {
  past: 'past',
  present: 'present',
  future: 'future',
} as const;

export interface ProjectUpdate {
  title?: string;
  description?: string;
  phase?: ProjectUpdatePhase;
  location?: string;
  style?: string;
  area?: string;
  imageUrl?: string;
  completionYear?: number;
  clientName?: string;
  featured?: boolean;
}

export type CustomerStatus = typeof CustomerStatus[keyof typeof CustomerStatus];


export const CustomerStatus = {
  hot: 'hot',
  warm: 'warm',
  cold: 'cold',
  past: 'past',
} as const;

export interface Customer {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  status: CustomerStatus;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  projectType?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  source?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type CustomerInputStatus = typeof CustomerInputStatus[keyof typeof CustomerInputStatus];


export const CustomerInputStatus = {
  hot: 'hot',
  warm: 'warm',
  cold: 'cold',
  past: 'past',
} as const;

export interface CustomerInput {
  name: string;
  email: string;
  phone?: string;
  status: CustomerInputStatus;
  budget?: string;
  projectType?: string;
  location?: string;
  notes?: string;
  source?: string;
}

export type CustomerUpdateStatus = typeof CustomerUpdateStatus[keyof typeof CustomerUpdateStatus];


export const CustomerUpdateStatus = {
  hot: 'hot',
  warm: 'warm',
  cold: 'cold',
  past: 'past',
} as const;

export interface CustomerUpdate {
  name?: string;
  email?: string;
  phone?: string;
  status?: CustomerUpdateStatus;
  budget?: string;
  projectType?: string;
  location?: string;
  notes?: string;
  source?: string;
}

export type InquiryStatus = typeof InquiryStatus[keyof typeof InquiryStatus];


export const InquiryStatus = {
  new: 'new',
  contacted: 'contacted',
  converted: 'converted',
  closed: 'closed',
} as const;

export interface Inquiry {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  message: string;
  /** @nullable */
  projectType?: string | null;
  /** @nullable */
  budget?: string | null;
  status: InquiryStatus;
  createdAt: string;
}

export interface InquiryInput {
  name: string;
  email: string;
  phone?: string;
  message: string;
  projectType?: string;
  budget?: string;
}

export type InquiryUpdateStatus = typeof InquiryUpdateStatus[keyof typeof InquiryUpdateStatus];


export const InquiryUpdateStatus = {
  new: 'new',
  contacted: 'contacted',
  converted: 'converted',
  closed: 'closed',
} as const;

export interface InquiryUpdate {
  status?: InquiryUpdateStatus;
  notes?: string;
}

export interface DashboardStats {
  totalCustomers: number;
  hotCount: number;
  warmCount: number;
  coldCount: number;
  pastCount: number;
  totalProjects: number;
  activeProjects: number;
  totalInquiries: number;
  newInquiries: number;
}

export type ListProjectsParams = {
phase?: ListProjectsPhase;
};

export type ListProjectsPhase = typeof ListProjectsPhase[keyof typeof ListProjectsPhase];


export const ListProjectsPhase = {
  past: 'past',
  present: 'present',
  future: 'future',
} as const;

export type ListCustomersParams = {
status?: ListCustomersStatus;
};

export type ListCustomersStatus = typeof ListCustomersStatus[keyof typeof ListCustomersStatus];


export const ListCustomersStatus = {
  hot: 'hot',
  warm: 'warm',
  cold: 'cold',
  past: 'past',
} as const;