# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Interior Design Studio
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Interior design projects
  - name: customers
    description: CRM customer management
  - name: inquiries
    description: Client inquiries
  - name: stats
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: phase
          in: query
          schema:
            type: string
            enum: [past, present, future]
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [hot, warm, cold, past]
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /inquiries:
    get:
      operationId: listInquiries
      tags: [inquiries]
      summary: List all inquiries
      responses:
        "200":
          description: List of inquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inquiry"
    post:
      operationId: createInquiry
      tags: [inquiries]
      summary: Submit an inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"

  /inquiries/{id}:
    patch:
      operationId: updateInquiry
      tags: [inquiries]
      summary: Update inquiry status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
    delete:
      operationId: deleteInquiry
      tags: [inquiries]
      summary: Delete an inquiry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /stats/dashboard:
    get:
      operationId: getDashboardStats
      tags: [stats]
      summary: Get CRM dashboard stats
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Project:
      type: object
      required: [id, title, phase, location, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        phase:
          type: string
          enum: [past, present, future]
        location:
          type: string
        style:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        completionYear:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        featured:
          type: boolean
        createdAt:
          type: string

    ProjectInput:
      type: object
      required: [title, phase, location]
      properties:
        title:
          type: string
        description:
          type: string
        phase:
          type: string
          enum: [past, present, future]
        location:
          type: string
        style:
          type: string
        area:
          type: string
        imageUrl:
          type: string
        completionYear:
          type: integer
        clientName:
          type: string
        featured:
          type: boolean

    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        phase:
          type: string
          enum: [past, present, future]
        location:
          type: string
        style:
          type: string
        area:
          type: string
        imageUrl:
          type: string
        completionYear:
          type: integer
        clientName:
          type: string
        featured:
          type: boolean

    Customer:
      type: object
      required: [id, name, email, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        status:
          type: string
          enum: [hot, warm, cold, past]
        budget:
          type: ["string", "null"]
        projectType:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    CustomerInput:
      type: object
      required: [name, email, status]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [hot, warm, cold, past]
        budget:
          type: string
        projectType:
          type: string
        location:
          type: string
        notes:
          type: string
        source:
          type: string

    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [hot, warm, cold, past]
        budget:
          type: string
        projectType:
          type: string
        location:
          type: string
        notes:
          type: string
        source:
          type: string

    Inquiry:
      type: object
      required: [id, name, email, message, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        message:
          type: string
        projectType:
          type: ["string", "null"]
        budget:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, contacted, converted, closed]
        createdAt:
          type: string

    InquiryInput:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        message:
          type: string
        projectType:
          type: string
        budget:
          type: string

    InquiryUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, contacted, converted, closed]
        notes:
          type: string

    DashboardStats:
      type: object
      required: [totalCustomers, hotCount, warmCount, coldCount, pastCount, totalProjects, activeProjects, totalInquiries, newInquiries]
      properties:
        totalCustomers:
          type: integer
        hotCount:
          type: integer
        warmCount:
          type: integer
        coldCount:
          type: integer
        pastCount:
          type: integer
        totalProjects:
          type: integer
        activeProjects:
          type: integer
        totalInquiries:
          type: integer
        newInquiries:
          type: integer