# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Interior Design Studio
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerStatus } from './customerStatus';

export interface Customer {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  status: CustomerStatus;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  projectType?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  source?: string | null;
  createdAt: string;
  updatedAt?: string;
}