# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Interior Design Studio
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerUpdateStatus } from './customerUpdateStatus';

export interface CustomerUpdate {
  name?: string;
  email?: string;
  phone?: string;
  status?: CustomerUpdateStatus;
  budget?: string;
  projectType?: string;
  location?: string;
  notes?: string;
  source?: string;
}